\name{convert}
\alias{convert}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
convert
}
\description{
This function to convert values between various units in five main areas such length, mass, time, temperature, data storage.
}
\usage{
length(value, from_unit, to_unit)
mass(value, from_unit, to_unit)
time(value, from_unit, to_unit)
temp(value, from_unit, to_unit)
storage(value, from_unit, to_unit)
}

\author{
Kelompok 6 Komstat
}
